pyproject.toml
[project]
name = "travel-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit",
    "openai-agents",
    "openai",
    "python-dotenv"
]
